+++ EasySaver.TextFile/EasySaver.TextFile.cs
@@ -44,6 +44,72 @@ namespace EasySaver.TextFile
 
         #endregion Write
 
+        #region Default Choices
+
+        // Default value if FileName is not provided.
+        private readonly static string _defaultFileName = "";
+
+        // Default value if NamingFormat is not provided.
+        private readonly static NamingFormat _defaultNamingFormat = NamingFormat.DateTime;
+
+        // Default value when using Save(string text, string fileName) and SaveToFolder(string text, string folderName, string fileName). Since there is only NamingFormat option available to use by design, this value is internally provided for shorter method usage.
+        private readonly static NamingFormat _OnlyTextProvidedNamingFormat = NamingFormat.Custom;
+
+        private readonly static string _defaultFolderName = "Data";
+        private readonly static bool _defaultOverwrite = false;
+        private readonly static bool _defaultRenameIfExists = true;
+
+        #endregion Default Choices
+
+        #region Save to default path
+
+        /// <summary>
+        /// Save file with into path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <returns>True or false.</returns>
+        public static bool Save(string text)
+        {
+            return Save(text: text, fileName: _defaultFileName, _defaultNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save file with into path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <returns>True or false.</returns>
+        public static bool Save(string text, string fileName)
+        {
+            return Save(text: text, fileName: fileName, _OnlyTextProvidedNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save file with into path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <returns>True or false.</returns>
+        public static bool Save(string text, string fileName, NamingFormat namingFormat)
+        {
+            return Save(text: text, fileName: fileName, namingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save file with into path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
+        /// <returns>True or false.</returns>
+        /// <returns></returns>
+        public static bool Save(string text, string fileName, NamingFormat namingFormat, bool overwrite)
+        {
+            return Save(text: text, fileName: fileName, namingFormat: namingFormat, overwrite: overwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
         /// <summary>
         /// Save file with into path with naming.
         /// </summary>
@@ -53,7 +119,7 @@ namespace EasySaver.TextFile
         /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
         /// <param name="renameIfExists">Specify if file should renamed with number if exists already.</param>
         /// <returns>True or false.</returns>
-        public static bool Save(string text, string fileName, NamingFormat namingFormat = NamingFormat.DateTime, bool overwrite = false, bool renameIfExists = true)
+        public static bool Save(string text, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)
         {
             // Calls GetFileName() to decide fileName with adding extension at the end.
             fileName = GetFileName(fileName: fileName, namingFormat: namingFormat);
@@ -81,7 +147,7 @@ namespace EasySaver.TextFile
                     {
                         // Saving file via WriteViaStreamWriter().
                         WriteViaStreamWriter(path: $"./{fileName}({i}){_defaultTextExtension}", text: text);
-                        
+
                         // Returning true as successful process.
                         return true;
                     }
@@ -100,6 +166,65 @@ namespace EasySaver.TextFile
             }
         }
 
+        #endregion Save to default path.
+
+        #region Save to folder path
+
+        /// <summary>
+        /// Save text into given folder path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        public static bool SaveToFolder(string text)
+        {
+            return SaveToFolder(text: text, folderName: _defaultFolderName, fileName: _defaultFileName, namingFormat: _defaultNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        public static bool SaveToFolder(string text, string folderName)
+        {
+            return SaveToFolder(text: text, folderName: folderName, fileName: _defaultFileName, namingFormat: _defaultNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        public static bool SaveToFolder(string text, string folderName, string fileName)
+        {
+            return SaveToFolder(text: text, folderName: folderName, fileName: fileName, NamingFormat.Custom, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        public static bool SaveToFolder(string text, string folderName, string fileName, NamingFormat namingFormat)
+        {
+            return SaveToFolder(text: text, fileName: fileName, folderName: folderName, namingFormat: namingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
+        public static bool SaveToFolder(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite)
+        {
+            return SaveToFolder(text: text, fileName: fileName, folderName: folderName, namingFormat: namingFormat, overwrite: overwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
         /// <summary>
         /// Save text into given folder path with naming.
         /// </summary>
@@ -109,8 +234,7 @@ namespace EasySaver.TextFile
         /// <param name="namingFormat">NamingFormat option to apply.</param>
         /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
         /// <param name="renameIfExists">Specify if file should renamed with number if exists already.</param>
-        /// <returns>True or false.</returns>
-        public static bool Save(string text, string fileName, string folderName = "Data", NamingFormat namingFormat = NamingFormat.DateTime, bool overwrite = false, bool renameIfExists = true)
+        public static bool SaveToFolder(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)
         {
             // Calls GetFileName() to decide fileName.
             fileName = GetFileName(fileName: fileName, namingFormat: namingFormat);
@@ -164,6 +288,71 @@ namespace EasySaver.TextFile
             }
         }
 
+        #endregion Save to folder path
+
+        /// <summary>
+        /// Save text into given folder path with naming.
+        /// </summary>
+        /// <param name="text">Text file to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
+        /// <param name="renameIfExists">Specify if file should renamed with number if exists already.</param>
+        [Obsolete(message: "This method is deprecated. Use SaveToFolder() instead.")]
+        public static bool Save(string text, string fileName, string folderName = "Data", NamingFormat namingFormat = NamingFormat.DateTime, bool overwrite = false, bool renameIfExists = true)
+        {
+            return SaveToFolder(text: text, folderName: folderName, fileName: fileName, namingFormat: namingFormat, overwrite: overwrite, renameIfExists: renameIfExists);
+        }
+
+        #region Save to default path (Safe)
+
+        /// <summary>
+        /// Save text file default folder with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveSafe(string text)
+        {
+            return SaveSafe(text: text, fileName: _defaultFileName, namingFormat: _defaultNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text file default folder with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveSafe(string text, string fileName)
+        {
+            return SaveSafe(text: text, fileName: fileName, namingFormat: _OnlyTextProvidedNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text file default folder with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveSafe(string text, string fileName, NamingFormat namingFormat)
+        {
+            return SaveSafe(text: text, fileName: fileName, namingFormat: namingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text file default folder with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveSafe(string text, string fileName, NamingFormat namingFormat, bool overwrite)
+        {
+            return SaveSafe(text: text, fileName: fileName, namingFormat: namingFormat, overwrite: overwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
         /// <summary>
         /// Save text file default folder with try-catch mechanism.
         /// </summary>
@@ -173,7 +362,7 @@ namespace EasySaver.TextFile
         /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
         /// <param name="renameIfExists">Specify if file should renamed with number if exists already.</param>
         /// <returns>True or false with exception.</returns>
-        public static (bool, Exception?) SaveSafe(string text, string fileName, NamingFormat namingFormat = NamingFormat.DateTime, bool overwrite = false, bool renameIfExists = true)
+        public static (bool, Exception?) SaveSafe(string text, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)
         {
             // Calls GetFileName() to decide fileName.
             fileName = GetFileName(fileName: fileName, namingFormat: namingFormat);
@@ -200,10 +389,10 @@ namespace EasySaver.TextFile
                     if (CheckIfFileExist(path: $"{fileName}({i}){_defaultTextExtension}") == false)
                     {
                         // Saving file via WriteViaStreamWriterSafe().
-                        (bool, Exception?) _result = WriteViaStreamWriterSafe(path: $"./{fileName}({i}{_defaultTextExtension}", text: text);
+                        (bool, Exception?) result = WriteViaStreamWriterSafe(path: $"./{fileName}({i}{_defaultTextExtension}", text: text);
 
                         // Returning result to indicate if process is successful or failed with exception.
-                        return _result;
+                        return result;
                     }
                 }
 
@@ -214,11 +403,75 @@ namespace EasySaver.TextFile
             else
             {
                 // Saving file via WriteViaStreamWriterSafe().
-                (bool, Exception?) _result = WriteViaStreamWriterSafe(path: $"./{fileName}{_defaultTextExtension}", text: text);
+                (bool, Exception?) result = WriteViaStreamWriterSafe(path: $"./{fileName}{_defaultTextExtension}", text: text);
 
                 // Returning result to indicate if process is successful or failed with exception.
-                return _result;
+                return result;
             }
+        }
+
+        #endregion Save to default path (Safe)
+
+        #region Save to folder path (Safe)
+
+        /// <summary>
+        /// Save text into given folder path with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveToFolderSafe(string text)
+        {
+            return SaveToFolderSafe(text: text, folderName: _defaultFolderName, fileName: _defaultFileName, namingFormat: _defaultNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveToFolderSafe(string text, string folderName)
+        {
+            return SaveToFolderSafe(text: text, folderName: folderName, fileName: _defaultFileName, namingFormat: _defaultNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveToFolderSafe(string text, string folderName, string fileName)
+        {
+            return SaveToFolderSafe(text: text, folderName: folderName, fileName: fileName, namingFormat: _OnlyTextProvidedNamingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveToFolderSafe(string text, string folderName, string fileName, NamingFormat namingFormat)
+        {
+            return SaveToFolderSafe(text: text, folderName: folderName, fileName: fileName, namingFormat: namingFormat, overwrite: _defaultOverwrite, renameIfExists: _defaultRenameIfExists);
+        }
+
+        /// <summary>
+        /// Save text into given folder path with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
+        /// <returns>True or false with exception.</returns>
+        public static (bool, Exception?) SaveToFolderSafe(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite)
+        {
+            return SaveToFolderSafe(text: text, folderName: folderName, fileName: fileName, namingFormat: namingFormat, overwrite: overwrite, renameIfExists: _defaultRenameIfExists);
         }
 
         /// <summary>
@@ -231,9 +484,9 @@ namespace EasySaver.TextFile
         /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
         /// <param name="renameIfExists">Specify if file should renamed with number if exists already.</param>
         /// <returns>True or false with exception.</returns>
-        public static (bool, Exception?) SaveSafe(string text, string fileName, string folderName = "Data", NamingFormat namingFormat = NamingFormat.DateTime, bool overwrite = false, bool renameIfExists = true)
+        public static (bool, Exception?) SaveToFolderSafe(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)
         {
-            //// Calls GetFileName() to decide fileName.
+            // Calls GetFileName() to decide fileName.
             fileName = GetFileName(fileName: fileName, namingFormat: namingFormat);
 
             // Checking if folder exists.
@@ -283,6 +536,25 @@ namespace EasySaver.TextFile
                 // Returning result to indicate if process is successful or failed with exception.
                 return result;
             }
+        }
+
+        #endregion Save to folder path (Safe)
+
+        /// <summary>
+        /// Save text into given folder path with try-catch mechanism.
+        /// </summary>
+        /// <param name="text">Text to save.</param>
+        /// <param name="fileName">Name of file when NamingFormat is Custom.</param>
+        /// <param name="folderName">Folder name to create if doesn't exist and use for saving file.</param>
+        /// <param name="namingFormat">NamingFormat option to apply.</param>
+        /// <param name="overwrite">Specify if file should overwritten if file exists already.</param>
+        /// <param name="renameIfExists">Specify if file should renamed with number if exists already.</param>
+        /// <returns>True or false with exception.</returns>
+        [Obsolete("This method is deprecated. Please use SaveToFolderSafe() instead.")]
+        public static (bool, Exception?) SaveSafe(string text, string fileName, string folderName = "Data", NamingFormat namingFormat = NamingFormat.DateTime, bool overwrite = false, bool renameIfExists = true)
+        {
+            //
+            return SaveToFolderSafe(text: text, folderName: folderName, fileName: fileName, namingFormat: namingFormat, overwrite: overwrite, renameIfExists: renameIfExists);
         }
     }
 }