+++ EasySaver.TextFile/CHANGELOG.md
@@ -10,6 +10,44 @@
 ### Removed
 -->
 
+## [1.0.0-alpha.5]
+
+### Added
+* `Save(string text)` method added.
+* `Save(string text, string fileName)`method added.
+* `Save(string text, string fileName, NamingFormat namingFormat)` method added.
+* `Save(string text, string fileName, NamingFormat namingFormat, bool overwrite)` method added.
+
+* `SaveToFolder(string text)` method added.
+* `SaveToFolder(string text, string folderName)` method added.
+* `SaveToFolder(string text, string folderName, string fileName)`method added.
+* `SaveToFolder(string text, string folderName, string fileName, NamingFormat namingFormat)` method added.
+* `SaveToFolder(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite)` method added.
+* `SaveToFolder(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)` method added.
+
+* `SaveSafe(string text)` method added.
+* `SaveSafe(string text, string fileName)`method added.
+* `SaveSafe(string text, string fileName, NamingFormat namingFormat)` method added.
+* `SaveSafe(string text, string fileName, NamingFormat namingFormat, bool overwrite)` method added.
+
+* `SaveToFolderSafe(string text)` method added.
+* `SaveToFolderSafe(string text, string folderName)` method added.
+* `SaveToFolderSafe(string text, string folderName, string fileName)`method added.
+* `SaveToFolderSafe(string text, string folderName, string fileName, NamingFormat namingFormat)` method added.
+* `SaveToFolderSafe(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite)` method added.
+* `SaveToFolderSafe(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)` method added.
+
+* Default provided values are listed as properties now.
+
+### Changed
+* `Save(string text, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExist)` method's default parameters are removed. NamingFormat, overwrite, renameIfExist had default parameters NamingFormat.DateTime, false, true respectively.
+* `SaveSafe(string text, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExist)` method's default parameters are removed. NamingFormat, overwrite, renameIfExist had default parameters NamingFormat.DateTime, false, true respectively.
+* Typo fixed.
+
+### Removed
+* `Save(string text, string fileName, string folderName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)` method is deprecated. You can use `SaveToFolder(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)` instead.
+* `SaveSafe(string text, string fileName, string folderName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)` method is deprecated. You can use `SaveToFolderSafe(string text, string folderName, string fileName, NamingFormat namingFormat, bool overwrite, bool renameIfExists)` instead.
+
 ## [1.0.0-alpha.4]
 
 ### Added